+++ src/lib/storage/index/b_tree/b_tree_index_impl.hpp
@@ -27,7 +27,7 @@ class BaseBTreeIndexImpl {
   virtual ~BaseBTreeIndexImpl() = default;
 
   using Iterator = std::vector<ChunkOffset>::const_iterator;
-  virtual uint64_t memory_consumption() const = 0;
+  virtual size_t memory_consumption() const = 0;
   virtual Iterator lower_bound(const std::vector<AllTypeVariant>&) const = 0;
   virtual Iterator upper_bound(const std::vector<AllTypeVariant>&) const = 0;
   virtual Iterator cbegin() const = 0;
@@ -56,7 +56,7 @@ class BTreeIndexImpl : public BaseBTreeIndexImpl {
   BTreeIndexImpl(BTreeIndexImpl&&) = default;
   BTreeIndexImpl& operator=(BTreeIndexImpl&&) = default;
 
-  uint64_t memory_consumption() const override;
+  size_t memory_consumption() const override;
 
   Iterator lower_bound(DataType value) const;
   Iterator upper_bound(DataType value) const;
